requirements.txt
# Production dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
sqlalchemy==2.0.32
python-multipart==0.0.9
pydantic[email]==2.9.2
pydantic-settings==2.0.3
pymysql==1.1.1
python-dotenv==1.0.1
alembic==1.13.2
itsdangerous==2.2.0
python-json-logger==2.0.7
jinja2==3.1.4
fastapi-mail==1.4.1

# Authentication dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1  # Fixed version for passlib compatibility

# Development dependencies (install only for development)
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1
bandit>=1.7.5
pytest>=7.4.3
pytest-cov>=4.1.0
pre-commit>=3.5.0
pip-audit>=2.6.1
packaging>=23.0.0
httpx>=0.27.0
